+++ bootique-kotlin-example/src/main/resources/config.bq.kts
@@ -22,34 +22,31 @@ import io.bootique.jetty.server.ServerFactory
 import io.bootique.kotlin.config.logback.consoleAppender
 import io.bootique.kotlin.config.logback.logbackContextFactory
 import io.bootique.kotlin.config.logback.logger
-import io.bootique.kotlin.config.modules.config
 import io.bootique.kotlin.sample.TestConfig
 import io.bootique.logback.LogbackLevel
 import io.bootique.logback.appender.ConsoleTarget
 
-config {
-    addConfig("test" to TestConfig(
-        name = "Ruslan",
-        products = listOf(
-            "Ubuntu",
-            "DELL",
-            "OnePlus"
+addConfig("test", TestConfig(
+    name = "Ruslan",
+    products = listOf(
+        "Ubuntu",
+        "DELL",
+        "OnePlus"
+    )
+))
+addConfig("jetty", ServerFactory())
+addConfig("log", logbackContextFactory(
+    useLogbackConfig = false,
+    debugLogback = false,
+    level = LogbackLevel.warn,
+    loggers = mapOf(
+        logger(Bootique::class, LogbackLevel.error)
+    ),
+    appenders = listOf(
+        consoleAppender(
+            target = ConsoleTarget.stderr,
+            logFormat = "%t %-5p %c{1}: %m%n"
         )
-    ))
-    addConfig("jetty" to ServerFactory())
-    addConfig("log" to logbackContextFactory(
-        useLogbackConfig = false,
-        debugLogback = false,
-        level = LogbackLevel.warn,
-        loggers = mapOf(
-            logger(Bootique::class, LogbackLevel.error)
-        ),
-        appenders = listOf(
-            consoleAppender(
-                target = ConsoleTarget.stderr,
-                logFormat = "%t %-5p %c{1}: %m%n"
-            )
-        ),
-        logFormat = "[%d{dd/MMM/yyyy:HH:mm:ss}] %t %-5p %c{1}: %m%n"
-    ))
-}
+    ),
+    logFormat = "[%d{dd/MMM/yyyy:HH:mm:ss}] %t %-5p %c{1}: %m%n"
+))